:bug: fix multiple issues
